+++ include/mruby_integration/models/local_storage.hpp
@@ -1,0 +1,5 @@
+#pragma once
+#include "mruby.h"
+
+void
+append_models_LocalStorage(mrb_state*);