+++ include/platform_specific/web.hpp
@@ -1,5 +1,0 @@
-#pragma once
-#include "mruby.h"
-
-void
-append_platform_specific_web(mrb_state*);