+++ mrb_doc/models/browser.rb
@@ -26,8 +26,8 @@ module Browser
   # @return [nil]
   # @raise [Taylor::Platform::MethodCalledOnInvalidPlatformError] If run on the wrong platform.
   def self.main_loop=(method)
-    # mrb_set_main_loop
-    # src/web.cpp
+    # mrb_Browser_main_loop
+    # src/mruby_integration/models/browser.cpp
     nil
   end
 
@@ -36,8 +36,28 @@ module Browser
   # @return [nil]
   # @raise [Taylor::Platform::MethodCalledOnInvalidPlatformError] If run on the wrong platform.
   def self.cancel_main_loop
-    # mrb_cancel_main_loop
-    # src/web.cpp
+    # mrb_Browser_main_loop
+    # src/mruby_integration/models/browser.cpp
     nil
+  end
+
+  # Gets the attribute from the specified element. If it's a data attribute you
+  # can just pass 'data-blah'.
+  #
+  # @example Basic usage
+  #   # Given an element <div class="save-data" data-save-json="{...}"></div>
+  #
+  #   puts Browser.attribute_from_element(".save-data", "data-save-json")
+  #   #=> "{...}"
+  #
+  #
+  # @param selector [String]
+  # @param attribute [String]
+  # @return [String]
+  # @raise [Taylor::Platform::MethodCalledOnInvalidPlatformError] If run on non-web platforms.
+  def self.attribute_from_element(selector, attribute)
+    # mrb_Browser_main_loop
+    # src/mruby_integration/models/browser.cpp
+    "data"
   end
 end