+++ src/mruby_integration/models/browser.cpp
@@ -49,6 +49,43 @@ mrb_Browser_cancel_main_loop(mrb_state*, mrb_value)
 
   return mrb_nil_value();
 }
+
+EM_JS(char*,
+      js_get_attribute_from_element,
+      (const char* selectorPointer, const char* attributePointer),
+      {
+        // clang-format off
+        const selector = Module.UTF8ToString(selectorPointer);
+        let attribute = Module.UTF8ToString(attributePointer);
+        let value;
+
+        if (attribute.slice(0, 5) === 'data-') {
+          attribute = attribute.slice(5, attribute.length);
+          value = document.querySelector(selector).dataset[attribute];
+        } else {
+          value = document.querySelector(selector)[attribute];
+        }
+
+        value = (value || "").toString();
+
+        const byteCount = Module.lengthBytesUTF8(value) + 1;
+        const valuePointer = Module._malloc(byteCount);
+
+        Module.stringToUTF8(value, valuePointer, byteCount);
+
+        return valuePointer;
+        // clang-format on
+      });
+
+mrb_value
+mrb_Browser_attribute_from_element(mrb_state* mrb, mrb_value)
+{
+  char *selector, *attribute;
+  mrb_get_args(mrb, "zz", &selector, &attribute);
+
+  return mrb_str_new_cstr(mrb,
+                          js_get_attribute_from_element(selector, attribute));
+}
 #endif
 
 void
@@ -69,6 +106,11 @@ append_models_Browser(mrb_state* mrb)
                           "cancel_main_loop",
                           mrb_Browser_cancel_main_loop,
                           MRB_ARGS_NONE());
+  mrb_define_class_method(mrb,
+                          Browser_class,
+                          "attribute_from_element",
+                          mrb_Browser_attribute_from_element,
+                          MRB_ARGS_REQ(2));
 #endif
 
 #ifndef __EMSCRIPTEN__
@@ -80,6 +122,10 @@ append_models_Browser(mrb_state* mrb)
 
       def self.cancel_main_loop
         raise Taylor::Platform::MethodCalledOnInvalidPlatformError, 'Browser.cancel_main_loop is only available for Web exports'
+      end
+
+      def self.attribute_from_element(selector, attribute)
+        raise Taylor::Platform::MethodCalledOnInvalidPlatformError, "get_attribute_from_element is only available for Web exports"
       end
     end
   )");