+++ src/mruby_integration/models/local_storage.cpp
@@ -1,46 +1,10 @@
 #include "mruby.h"
-#include "mruby/compile.h"
 #include <mruby/class.h>
 
-#include "ruby/platform_specific/web.hpp"
+#include "ruby/models/local_storage.hpp"
 
 #ifdef __EMSCRIPTEN__
 #include <emscripten/emscripten.h>
-
-EM_JS(char*,
-      js_get_attribute_from_element,
-      (const char* selectorPointer, const char* attributePointer),
-      {
-        // clang-format off
-        const selector = Module.UTF8ToString(selectorPointer);
-        let attribute = Module.UTF8ToString(attributePointer);
-        let value;
-
-        if (attribute.slice(0, 5) === 'data-') {
-          attribute = attribute.slice(5, attribute.length);
-          value = document.querySelector(selector).dataset[attribute];
-        } else {
-          value = document.querySelector(selector)[attribute];
-        }
-
-        const byteCount = Module.lengthBytesUTF8(value) + 1;
-        const valuePointer = Module._malloc(byteCount);
-
-        Module.stringToUTF8(value, valuePointer, byteCount);
-
-        return valuePointer;
-        // clang-format on
-      });
-
-mrb_value
-mrb_get_attribute_from_element(mrb_state* mrb, mrb_value)
-{
-  char *selector, *attribute;
-  mrb_get_args(mrb, "zz", &selector, &attribute);
-
-  return mrb_str_new_cstr(mrb,
-                          js_get_attribute_from_element(selector, attribute));
-}
 
 EM_JS(void,
       js_local_storage_set_item,
@@ -93,7 +57,7 @@ mrb_local_storage_get_item(mrb_state* mrb, mrb_value)
 #endif
 
 void
-append_platform_specific_web(mrb_state* mrb)
+append_models_LocalStorage(mrb_state* mrb)
 {
 #ifdef __EMSCRIPTEN__
   struct RClass* LocalStorage_class;
@@ -109,15 +73,9 @@ append_platform_specific_web(mrb_state* mrb)
                           "get_item",
                           mrb_local_storage_get_item,
                           MRB_ARGS_REQ(2));
-
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "get_attribute_from_element",
-                    mrb_get_attribute_from_element,
-                    MRB_ARGS_REQ(2));
 #endif
 
 #ifndef __EMSCRIPTEN__
-  load_ruby_platform_specific_web(mrb);
+  load_ruby_models_local_storage(mrb);
 #endif
 }