+++ src/ruby/models/local_storage.rb
@@ -1,17 +1,8 @@
-# Gets the attribute from the specified element. If it's a data attribute you
-# can just pass 'data-blah'.
-# @param selector [String]
-# @param attribute [String]
-# @return [String]
-# @raise [Taylor::Platform::MethodCalledOnInvalidPlatformError] If run on non-web platforms.
-def get_attribute_from_element(selector, attribute)
-  raise Taylor::Platform::MethodCalledOnInvalidPlatformError, "get_attribute_from_element is only available for Web exports"
-end
-
-# A class for interacting with the browser's LocalStorage.
+# A class for interacting with the browser's localStorage.
 class LocalStorage
   # Returns the value stored in localStorage in the browser, if it doesn't exist
   # it will just return an empty string.
+  #
   # @param key [String]
   # @return [String]
   # @raise [Taylor::Platform::MethodCalledOnInvalidPlatformError] If run on non-web platforms.
@@ -20,6 +11,7 @@ class LocalStorage
   end
 
   # Sets the value in localStorage in the browser.
+  #
   # @param key [String]
   # @param value [String]
   # @return [nil]